move Schema into own folder@
