+++ website/src/pages/dictionary/Schema/styles.module.css
@@ -1,0 +1,24 @@
+.schemaTitle {
+  font-size: 24px;
+}
+
+.schemaTable_column {
+  max-width: 450px;
+  min-width: 100px;
+}
+
+.requiredTag {
+  background-color: #df1b42;
+  font-size: 12px;
+  color: white;
+  border-radius: 5px;
+  text-align: center;
+}
+
+.fieldDescription {
+  font-size: 12px;
+}
+
+.fieldEnumValue {
+  font-size: 12px;
+}
